# docs/swagger.yaml
basePath: /
definitions:
  domain.Address:
    properties:
      apartment:
        type: string
      door_code:
        type: string
      entrance:
        type: string
      floor:
        type: string
      id:
        type: integer
      street_house:
        type: string
      user_id:
        type: integer
    type: object
  domain.UpdateUserParams:
    properties:
      email:
        type: string
      name:
        type: string
      phone:
        type: string
    type: object
  domain.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      role:
        $ref: '#/definitions/domain.UserRole'
      updatedAt:
        type: string
    type: object
  domain.UserRole:
    enum:
    - user
    - admin
    type: string
    x-enum-varnames:
    - RoleUser
    - RoleAdmin
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  handlers.loginReq:
    properties:
      password:
        type: string
      phone:
        type: string
    type: object
  handlers.registerReq:
    properties:
      name:
        type: string
      password:
        type: string
      phone:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API для авторизации и профиля
  title: Mixfood Auth API
  version: "1.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Телефон и пароль
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.loginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Логин
      tags:
      - Auth
  /api/auth/logout:
    post:
      responses:
        "204":
          description: No Content
      summary: Выход
      tags:
      - Auth
  /api/auth/refresh:
    get:
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Обновление токенов
      tags:
      - Auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные пользователя
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.registerReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Регистрация
      tags:
      - Auth
  /api/user/address:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные адреса
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.Address'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Создать адрес
      tags:
      - User
  /api/user/addresses:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Address'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Получить мои адреса
      tags:
      - User
  /api/user/profile:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Данные
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateUserParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
      summary: Обновить профиль
      tags:
      - User
  /api/users/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Получить профиль пользователя
      tags:
      - Profile
swagger: "2.0"